+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `applicationFormId` is now a field on `bulkUploadTasks`
+
+
 ## 0.28.0 2025-01-08
 
 ### Added